+++ nextjs-app/src/app/home/page.tsx
@@ -9,6 +9,7 @@ import iconSeta from '../../assets/icons/arrow_back_ios_new.svg'
 import { useRouter } from "next/navigation";
 import { hasAccessToken } from "../../lib/auth/hasAccessToken";
 import { useEffect, useState } from "react";
+import Background from "@/components/Background";
 
 export default function HomePage() {
   const router = useRouter()
@@ -30,55 +31,57 @@ export default function HomePage() {
   }, [accessTokenExists])
 
   return (
-    <div className="flex min-w-full h-header">
-      <aside className="flex flex-col gap-3 px-9 pt-4 w-1/5 bg-neutral-1">
-        <span className="flex gap-4">
-          <Image alt="icon-inicio" src={iconInicio} />
-          <h5 className="text-h5 text-neutral-2">Início</h5>
-        </span>
-        <span className="flex gap-4">
-          <Image alt="icon-inicio" src={iconRecentes} />
-          <h5 className="text-h5 text-neutral-2">Recentes</h5>
-        </span>
-        <span className="flex gap-4">
-          <Image alt="icon-inicio" src={iconFavoritos} />
-          <h5 className="text-h5 text-neutral-2">Favoritos</h5>
-        </span>
-        <span className="flex gap-4">
-          <Image alt="icon-inicio" src={iconLixeira} />
-          <h5 className="text-h5 text-neutral-2">Lixeira</h5>
-        </span>
-        <span className="line my-2"></span>
-        <span className="flex gap-4">
-          <Image alt="icon-inicio" src={iconSeta} />
-          <h5 className="text-h5 text-neutral-2">Times</h5>
-        </span>
-      </aside>
-      <section className="flex flex-col w-4/5 pl-12 pr-20 ">
-        <h3 className="text-h3 py-4 text-neutral-2 font-weight-bold">Início</h3>
-        <div className="pb-6">
-          <h4 className="text-h4 text-neutral-2 pb-2">Acessar documento</h4>
-          <div className="flex gap-3">
-            <button className="filter-btn">Time</button>
-            <button className="filter-btn">Minha biblioteca</button>
-            <button className="filter-btn">Favoritos</button>
-          </div>
-        </div>
-        <div>
-          <div className="w-full bg-neutral-1 pl-8 py-6 pr-11 rounded-lg">
-            <h3 className="text-h3 mb-1 text-neutral-2 font-weight-bold">Documento teste</h3>
-            <div className="flex items-center gap-1 mb-3">
-              <span className="tag">#Python</span>
-              <span className="tag">#Dev</span>
-            </div>
-            <p className="text-p text-neutral-2 mb-4">Lorem ipsum dolor sit amet consectetur. Gravida ut gravida eu viverra eget mollis. Volutpat in cursus imperdiet est. Cursus tincidunt sem lacus facilisis pellentesque neque diam. Donec odio sit massa in fermentum nec tempus quam diam.Lorem ipsum dolor sit amet consectetur. Gravida ut gravida eu viverra eget mollis. Volutpat in cursus imperdiet est. Cursus tincidunt sem lacus facilisis pellentesque neque diam. Donec odio sit massa in fermentum nec tempus quam diam.</p>
-            <div className="flex justify-between w-full">
-              <span className="border border-primary-1 text-primary-1 px-4 py-1 rounded-lg">Time</span>
-              <button className="highlight-btn-sm">Acessar</button>
+    <Background custom="overflow-hidden">
+      <div className="flex min-w-full h-header">
+        <aside className="flex flex-col gap-3 px-9 pt-4 w-1/5 bg-neutral-1">
+          <span className="flex gap-4">
+            <Image alt="icon-inicio" src={iconInicio} />
+            <h5 className="text-h5 text-neutral-2">Início</h5>
+          </span>
+          <span className="flex gap-4">
+            <Image alt="icon-inicio" src={iconRecentes} />
+            <h5 className="text-h5 text-neutral-2">Recentes</h5>
+          </span>
+          <span className="flex gap-4">
+            <Image alt="icon-inicio" src={iconFavoritos} />
+            <h5 className="text-h5 text-neutral-2">Favoritos</h5>
+          </span>
+          <span className="flex gap-4">
+            <Image alt="icon-inicio" src={iconLixeira} />
+            <h5 className="text-h5 text-neutral-2">Lixeira</h5>
+          </span>
+          <span className="line my-2"></span>
+          <span className="flex gap-4">
+            <Image alt="icon-inicio" src={iconSeta} />
+            <h5 className="text-h5 text-neutral-2">Times</h5>
+          </span>
+        </aside>
+        <section className="flex flex-col w-4/5 pl-12 pr-20 ">
+          <h3 className="text-h3 py-4 text-neutral-2 font-weight-bold">Início</h3>
+          <div className="pb-6">
+            <h4 className="text-h4 text-neutral-2 pb-2">Acessar documento</h4>
+            <div className="flex gap-3">
+              <button className="filter-btn">Time</button>
+              <button className="filter-btn">Minha biblioteca</button>
+              <button className="filter-btn">Favoritos</button>
             </div>
           </div>
-        </div>
-      </section>
-    </div>
+          <div>
+            <div className="w-full bg-neutral-1 pl-8 py-6 pr-11 rounded-lg">
+              <h3 className="text-h3 mb-1 text-neutral-2 font-weight-bold">Documento teste</h3>
+              <div className="flex items-center gap-1 mb-3">
+                <span className="tag">#Python</span>
+                <span className="tag">#Dev</span>
+              </div>
+              <p className="text-p text-neutral-2 mb-4">Lorem ipsum dolor sit amet consectetur. Gravida ut gravida eu viverra eget mollis. Volutpat in cursus imperdiet est. Cursus tincidunt sem lacus facilisis pellentesque neque diam. Donec odio sit massa in fermentum nec tempus quam diam.Lorem ipsum dolor sit amet consectetur. Gravida ut gravida eu viverra eget mollis. Volutpat in cursus imperdiet est. Cursus tincidunt sem lacus facilisis pellentesque neque diam. Donec odio sit massa in fermentum nec tempus quam diam.</p>
+              <div className="flex justify-between w-full">
+                <span className="border border-primary-1 text-primary-1 px-4 py-1 rounded-lg">Time</span>
+                <button className="highlight-btn-sm">Acessar</button>
+              </div>
+            </div>
+          </div>
+        </section>
+      </div>
+    </Background>
   );
 }
