+++ nextjs-app/src/app/signup/components/ModalTerms.tsx
